refactor(actions): move the classifier embed adapter to embedder-swap
embedClassifierDescriptions resolves a config off the open story and picks a provider, which is composeRetrievalEmbedDeps' shape rather than a classifier-scheduler concern; sitting next to resolveDrainConfig it reuses providerFor instead of duplicating the lookup, and deps.ts loses its cross-folder embedder-swap import.
