+++ lib/actions/embedder-swap/index.ts
@@ -16,6 +16,7 @@ export {
   composeRetrievalEmbedDeps,
   kickStoryDrain,
   countStoryEmbeddableRows,
+  embedClassifierDescriptions,
   isStorySwapPending,
   refreshEmbeddingStatus,
   reindexStoryNow,
